Set foreign key referred schema to None when needed
SQLAlchemy expects foreign key reflection to set schemas which
do not need specifying in the current database context (i.e.
default schema or current schema) to be set to None.

Without this, alembic migration keep re-creating foreign keys
as it considers a referred table with a schema and without a
schema two separate tables.